refactor: add db and models packages, dyndict into a separate module
