+++ code/nuvla/notifs/models/resource.py
